url-el sem lehet elérni más user dolgait, feature tesztek zöldek + 3 teszt refaktorálása

+++ example-app/app/Http/Controllers/AlbumsController.php
@@ -117,7 +117,7 @@ class AlbumsController extends Controller
     public function destroy($id)
     {
         $album=Album::find ($id);
-        if(Storage::delete('/public/storage/albums/'.$album->album_id.'/'.$album->album)) {
+        if(Storage::delete('/public/storage/albums/'.$album->album_id.'/'.$album->album) && $album->ulby==auth()->id()) {
             $album->delete();
             return redirect('/albums')->with('success', 'Album deleted successfully');
         }
